Add Details page and Change DataType int To long

+++ eCommerceApp.Models/Stock.cs
@@ -7,9 +7,9 @@ namespace eCommerceApp.Models
 {
     public class Stock
     {
-        public int Id { get; set; }
-        public int TotalStock { get; set; }
-        public int ProductId { get; set; }
+        public long Id { get; set; }
+        public long TotalStock { get; set; }
+        public long ProductId { get; set; }
         public Product Product { get; set; }
     }
 }
